+++ app/lib/landings/index.ts
@@ -13,7 +13,7 @@ import { productivityListicle } from "./productivity-listicle";
 import { brainAgeingListicle } from "./brain-ageing-listicle";
 import { generalListicle } from "./general-listicle";
 import type { OfferConfig } from "./offer-types";
-import { flowTrial } from "./flow-trial";
+import { flowFourBox } from "./flow-4-box";
 
 /** Any landing page config; narrow on `format` to render */
 export type AnyLandingConfig = LandingConfig | ListicleConfig | OfferConfig;
@@ -29,7 +29,7 @@ const registry: Record<string, AnyLandingConfig> = {
   [brainAgeingListicle.slug]: brainAgeingListicle,
   [generalListicle.slug]: generalListicle,
   // Offer pages (SCRUM-1343): one config per offer, see offer-types.ts.
-  [flowTrial.slug]: flowTrial,
+  [flowFourBox.slug]: flowFourBox,
 };
 
 export const landingSlugs = Object.keys(registry);